+++ src/components/form.js
@@ -50,7 +50,15 @@
         const mounted = useRef(false);
 
         B.defineFunction('Submit', () => {
-          if (formRef.current) formRef.current.requestSubmit();
+          if (formRef.current) {
+            if (typeof formRef.current.requestSubmit === 'function') {
+              formRef.current.requestSubmit();
+            } else {
+              formRef.current.dispatchEvent(
+                new Event('submit', { cancelable: true }),
+              );
+            }
+          }
         });
 
         const [, setOptions] = useOptions();
@@ -134,6 +142,20 @@
           }
         };
 
+        const FormElement = (
+          <form
+            className={[
+              empty && classes.empty,
+              isPristine && classes.pristine,
+            ].join(' ')}
+          >
+            {isPristine && (
+              <span>Drag form components in the form to submit data</span>
+            )}
+            {children}
+          </form>
+        );
+
         const FormCmp = ({ item }) => {
           const [isInvalid, setIsInvalid] = useState(false);
           const handleInvalid = () => {
@@ -142,6 +164,7 @@
               B.triggerEvent('onInvalid');
             }
           };
+
           useEffect(() => {
             B.triggerEvent('onComponentRendered');
           }, []);
@@ -249,7 +272,9 @@
           return <FormCmp item={item} />;
         };
 
-        return hasFilter ? <FormWithData /> : <FormCmp />;
+        const RuntimeForm = hasFilter ? <FormWithData /> : <FormCmp />;
+
+        return isDev ? FormElement : RuntimeForm;
       })()}
     </div>
   ),