+++ src/prefabs/pageWithLoginAndRegisterForm.js
@@ -3048,6 +3048,40 @@
                                             },
                                           },
                                           {
+                                            value: 'default',
+                                            label: 'Show on load',
+                                            key: 'loadingType',
+                                            type: 'CUSTOM',
+                                            configuration: {
+                                              as: 'BUTTONGROUP',
+                                              dataType: 'string',
+                                              allowedInput: [
+                                                {
+                                                  name: 'Message',
+                                                  value: 'default',
+                                                },
+                                                {
+                                                  name: 'Content',
+                                                  value: 'showChildren',
+                                                },
+                                              ],
+                                            },
+                                          },
+                                          {
+                                            value: ['Loading...'],
+                                            label: 'Loading text',
+                                            key: 'loadingText',
+                                            type: 'VARIABLE',
+                                            configuration: {
+                                              condition: {
+                                                type: 'SHOW',
+                                                option: 'loadingType',
+                                                comparator: 'EQ',
+                                                value: 'default',
+                                              },
+                                            },
+                                          },
+                                          {
                                             value: [
                                               '0rem',
                                               '0rem',
@@ -4775,6 +4809,40 @@
                                                 option: 'showError',
                                                 comparator: 'EQ',
                                                 value: 'built-in',
+                                              },
+                                            },
+                                          },
+                                          {
+                                            value: 'default',
+                                            label: 'Show on load',
+                                            key: 'loadingType',
+                                            type: 'CUSTOM',
+                                            configuration: {
+                                              as: 'BUTTONGROUP',
+                                              dataType: 'string',
+                                              allowedInput: [
+                                                {
+                                                  name: 'Message',
+                                                  value: 'default',
+                                                },
+                                                {
+                                                  name: 'Content',
+                                                  value: 'showChildren',
+                                                },
+                                              ],
+                                            },
+                                          },
+                                          {
+                                            value: ['Loading...'],
+                                            label: 'Loading text',
+                                            key: 'loadingText',
+                                            type: 'VARIABLE',
+                                            configuration: {
+                                              condition: {
+                                                type: 'SHOW',
+                                                option: 'loadingType',
+                                                comparator: 'EQ',
+                                                value: 'default',
                                               },
                                             },
                                           },